pyproject.toml
[project]
name = "email-intel"
version = "0.1.0"
description = "Local autonomous email intelligence assistant"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "tenacity>=8.3",
    "beautifulsoup4>=4.12",
    "html2text>=2024.2.26",
    "dateparser>=1.2",
    "imap-tools>=1.6",
    "python-dotenv>=1.0",
    "google-api-python-client>=2.130",
    "google-auth>=2.30",
    "google-auth-oauthlib>=1.2",
]

[project.scripts]
email-intel = "email_intel.__main__:main"
email-intel-once = "email_intel.app:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/email_intel"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["email_intel"]
mypy_path = "src"
